+++ test/object-store/realm.cpp
@@ -502,6 +502,33 @@ TEST_CASE("SharedRealm: get_shared_realm()") {
         REQUIRE(realm->read_transaction_version() == frozen2->read_transaction_version());
         REQUIRE(frozen2->read_transaction_version() > frozen1->read_transaction_version());
     }
+
+    SECTION("frozen realm should have the same schema as originating realm") {
+        auto full_schema = Schema{
+            {"object1", {{"value", PropertyType::Int}}},
+            {"object2", {{"value", PropertyType::Int}}},
+        };
+
+        auto subset_schema = Schema{
+            {"object1", {{"value", PropertyType::Int}}},
+        };
+
+        config.schema = full_schema;
+
+        auto realm = Realm::get_shared_realm(config);
+        realm->close();
+
+        config.schema = subset_schema;
+
+        realm = Realm::get_shared_realm(config);
+        realm->read_group();
+        auto frozen_realm = realm->freeze();
+        auto frozen_schema = frozen_realm->schema();
+
+        REQUIRE(full_schema != subset_schema);
+        REQUIRE(realm->schema() == subset_schema);
+        REQUIRE(frozen_schema == subset_schema);
+    }
 }
 
 #if REALM_ENABLE_SYNC
@@ -1068,11 +1095,7 @@ TEST_CASE("Realm::delete_files()") {
     // We need to create some additional files that might not be present
     // for a freshly opened realm but need to be tested for as the will
     // be created during a Realm's life cycle.
-    // .log_a and .log_b are legacy versions of .log that might be present
-    // for older Realms.
     util::File(path + ".log", util::File::mode_Write);
-    util::File(path + ".log_a", util::File::mode_Write);
-    util::File(path + ".log_b", util::File::mode_Write);
 
     SECTION("Deleting files of a closed Realm succeeds.") {
         realm->close();
@@ -1083,8 +1106,6 @@ TEST_CASE("Realm::delete_files()") {
         REQUIRE_FALSE(util::File::exists(path + ".management"));
         REQUIRE_FALSE(util::File::exists(path + ".note"));
         REQUIRE_FALSE(util::File::exists(path + ".log"));
-        REQUIRE_FALSE(util::File::exists(path + ".log_a"));
-        REQUIRE_FALSE(util::File::exists(path + ".log_b"));
 
         // Deleting the .lock file is not safe. It must still exist.
         REQUIRE(util::File::exists(path + ".lock"));
@@ -1097,8 +1118,6 @@ TEST_CASE("Realm::delete_files()") {
         REQUIRE(util::File::exists(path + ".management"));
         REQUIRE(util::File::exists(path + ".note"));
         REQUIRE(util::File::exists(path + ".log"));
-        REQUIRE(util::File::exists(path + ".log_a"));
-        REQUIRE(util::File::exists(path + ".log_b"));
     }
 
     SECTION("Deleting the same Realm multiple times.") {