+++ test/test_sync.cpp
@@ -5917,8 +5917,6 @@ TEST(Sync_ClientErrorHandler)
 }
 
 
-#ifndef REALM_PLATFORM_WIN32
-
 TEST(Sync_VerifyServerHistoryAfterLargeUpload)
 {
     TEST_DIR(server_dir);
@@ -5959,8 +5957,6 @@ TEST(Sync_VerifyServerHistoryAfterLargeUpload)
         }
     }
 }
-
-#endif // REALM_PLATFORM_WIN32
 
 
 TEST(Sync_ServerSideModify_Randomize)